+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 1.0.0 ##
+### 1.0.1 -- 2013 April 23 ###
 
-• first release
+- added an optional "cost limit" to `TMMemoryCache`, including new properties and methods
+- calling `[TMDiskCache trimToDate:]` with `[NSDate distantPast]` will now clear the cache
+- setting the byte limit on `TMDiskCache` to 0 will no longer clear the cache (0 means no limit)